pyproject.toml
[tool.poetry]
name = "mephi-ifib"
version = "0.1.0"
description = ""
authors = ["arsenyvolodko <arvolodko@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.2.0"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-camel-case = "^1.4.2"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
django-cors-headers = "^4.4.0"
inflection = "^0.5.1"
celery = "^5.4.0"
redis = "^5.2.0"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
